Moved assignment of camera to viewport
